Rename test files to fit naming convention
